# docs/api-server/openapi.yaml
openapi: "3.1.0"
info:
  title: AgentClash API
  version: 1.0.0
  description: >
    B2B AI agent evaluation platform. Pits agent deployments against challenge
    packs, with replay inspection, scoring, comparison, and release gating.
  license:
    name: MIT
  contact:
    url: https://github.com/Atharva-Kanherkar/agentclash
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/auth/device:
    post:
      summary: Create device authorization code
      description: >
        Initiates the device code flow (RFC 8628). Returns a device_code (secret)
        and user_code (displayed to user). The CLI polls /v1/auth/device/token
        with the device_code until the user approves.
      tags: [CLIAuth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Device code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceCodeResponse"
  /v1/auth/device/token:
    post:
      summary: Poll for device token
      description: >
        CLI polls this endpoint with the device_code until the user approves,
        denies, or the code expires. Returns authorization_pending, access_denied,
        expired_token, or the CLI token on approval.
      tags: [CLIAuth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_code]
              properties:
                device_code:
                  type: string
      responses:
        "200":
          description: Device approved, token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollDeviceTokenResponse"
        "400":
          description: Pending, denied, or expired
...
components:
...
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT Bearer token. In development mode, use custom headers instead:
        X-Agentclash-User-Id, X-Agentclash-WorkOS-User-Id,
        X-Agentclash-User-Email, X-Agentclash-User-Display-Name,
        X-Agentclash-Workspace-Memberships.
    hostedRunToken:
      type: http
      scheme: bearer
      description: >
        HMAC-SHA256 callback token issued when a hosted run is started.
        Not a JWT — the token encodes run_id and run_agent_id claims.
    artifactSignature:
      type: apiKey
      in: query
      name: signature
      description: >
        HMAC-SHA256 signature for artifact content URLs. Must be paired with
        an "expires" query parameter (unix timestamp).

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
...
  schemas:
...
    DeviceCodeResponse:
      type: object
      required: [device_code, user_code, verification_uri, expires_in, interval]
      properties:
        device_code:
          type: string
        user_code:
          type: string
          example: "RRGQ-BJVS"
        verification_uri:
          type: string
        expires_in:
          type: integer
        interval:
          type: integer
    PollDeviceTokenResponse:
      type: object
      properties:
        token:
          type: string
        user_id:
          type: string
        email:
          type: string
...